+++ hata/discord/soundboard/soundboard_sound/utils.py
@@ -2,8 +2,18 @@ __all__ = ('create_partial_soundboard_sound_from_id', 'create_partial_soundboard
 
 from ...core import SOUNDBOARD_SOUNDS
 
-from .fields import parse_guild_id, parse_id
+from .fields import (
+    parse_guild_id, parse_id, put_emoji_into, put_name_into, put_volume_into, validate_emoji, validate_name,
+    validate_volume
+)
 from .soundboard_sound import SoundboardSound
+
+
+SOUNDBOARD_SOUND_FIELD_CONVERTERS = {
+    'emoji': (validate_emoji, put_emoji_into),
+    'name': (validate_name, put_name_into),
+    'volume': (validate_volume, put_volume_into),
+}
 
 
 def create_partial_soundboard_sound_from_id(sound_id, guild_id = 0):